accounts: Redirect to subdomain'd url if logged in
Note that this is fixes a different bug than the trello one.
Trello bug :
Go to hyrodactil.com:8000 while logged in, it doesn't redirect you to
the subdomain and you get a 404 when clicking on a menu

This is due to django-debug-toolbar re-loading the page twice.

This fixes the bug where the user already logged in would access
hyrodactil.com without subdomain and should be redirected to its home
page. This accidentally also fixes the trello bug.